modified correct index, to show the right answer out of the questions answered, instead of all the questions

+++ PinYinPracticeForiPhone/ListenItemViewController.h
@@ -20,6 +20,6 @@
 @property (nonatomic, strong) UIButton *autoPlayButton;
 @property (nonatomic, strong) UILabel *indexLabel;
 
-@property (nonatomic, assign) NSInteger tag;
+@property (nonatomic, assign) int tag;
 
 @end
